+++ clevertap-core/src/main/java/com/clevertap/android/sdk/InAppNotificationActivity.java
@@ -1,14 +1,27 @@
 package com.clevertap.android.sdk;
 
+import static com.clevertap.android.sdk.CTXtensions.isPackageAndOsTargetsAbove;
+import static com.clevertap.android.sdk.inapp.InAppController.DISPLAY_HARD_PERMISSION_BUNDLE_KEY;
+import static com.clevertap.android.sdk.inapp.InAppController.IS_FIRST_TIME_PERMISSION_REQUEST;
+import static com.clevertap.android.sdk.inapp.InAppController.SHOW_FALLBACK_SETTINGS_BUNDLE_KEY;
+
+import android.Manifest;
+import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.WindowManager;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentActivity;
+import com.clevertap.android.sdk.inapp.AlertDialogPromptForSettings;
 import com.clevertap.android.sdk.inapp.CTInAppBaseFullFragment;
 import com.clevertap.android.sdk.inapp.CTInAppHtmlCoverFragment;
 import com.clevertap.android.sdk.inapp.CTInAppHtmlHalfInterstitialFragment;
@@ -20,10 +33,12 @@ import com.clevertap.android.sdk.inapp.CTInAppNativeHalfInterstitialImageFragmen
 import com.clevertap.android.sdk.inapp.CTInAppNativeInterstitialFragment;
 import com.clevertap.android.sdk.inapp.CTInAppNativeInterstitialImageFragment;
 import com.clevertap.android.sdk.inapp.CTInAppNotification;
+import com.clevertap.android.sdk.inapp.CTInAppNotificationButton;
 import com.clevertap.android.sdk.inapp.CTInAppType;
 import com.clevertap.android.sdk.inapp.InAppListener;
 import java.lang.ref.WeakReference;
 import java.util.HashMap;
+import kotlin.Unit;
 
 public final class InAppNotificationActivity extends FragmentActivity implements InAppListener {
 
@@ -34,6 +49,23 @@ public final class InAppNotificationActivity extends FragmentActivity implements
     private CTInAppNotification inAppNotification;
 
     private WeakReference<InAppListener> listenerWeakReference;
+
+    private WeakReference<PermissionCallback> permissionCallbackWeakReference;
+
+    private static final int PERMISSION_REQUEST_CODE = 2;
+
+    public static final String ANDROID_PERMISSION_STRING = "android.permission.POST_NOTIFICATIONS";
+
+    private boolean isFallbackSettingsEnabled;
+
+    private boolean shouldShowFallbackSettings;
+
+    public interface PermissionCallback {
+
+        void onAccept();
+
+        void onReject();
+    }
 
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -46,12 +78,24 @@ public final class InAppNotificationActivity extends FragmentActivity implements
             if (notif == null) {
                 throw new IllegalArgumentException();
             }
-            inAppNotification = notif.getParcelable("inApp");
+            inAppNotification = notif.getParcelable(Constants.INAPP_KEY);
+            boolean showHardNotificationPermission = notif.getBoolean(DISPLAY_HARD_PERMISSION_BUNDLE_KEY,
+                    false); // Using this boolean for a directly showing hard permission dialog flow
             Bundle configBundle = notif.getBundle("configBundle");
             if (configBundle != null) {
                 config = configBundle.getParcelable("config");
             }
             setListener(CleverTapAPI.instanceWithConfig(this, config).getCoreState().getInAppController());
+
+            setPermissionCallback(CleverTapAPI.instanceWithConfig(this, config).getCoreState()
+                    .getInAppController());
+
+            if (showHardNotificationPermission) {
+                shouldShowFallbackSettings = notif.getBoolean(SHOW_FALLBACK_SETTINGS_BUNDLE_KEY,
+                        false);
+                showHardPermissionPrompt();
+                return;
+            }
         } catch (Throwable t) {
             Logger.v("Cannot find a valid notification bundle to show!", t);
             finish();
@@ -111,8 +155,8 @@ public final class InAppNotificationActivity extends FragmentActivity implements
 
     @Override
     public void inAppNotificationDidClick(CTInAppNotification inAppNotification, Bundle formData,
-            HashMap<String, String> keyValueMap) {
-        didClick(formData, keyValueMap);
+            HashMap<String, String> keyValueMap, int btnClickIndex) {
+        didClick(formData, keyValueMap, btnClickIndex);
     }
 
     @Override
@@ -139,11 +183,107 @@ public final class InAppNotificationActivity extends FragmentActivity implements
         super.setTheme(android.R.style.Theme_Translucent_NoTitleBar);
     }
 
-    void didClick(Bundle data, HashMap<String, String> keyValueMap) {
+    void didClick(Bundle data, HashMap<String, String> keyValueMap, int btnClickIndex) {
         InAppListener listener = getListener();
         if (listener != null) {
-            listener.inAppNotificationDidClick(inAppNotification, data, keyValueMap);
+            listener.inAppNotificationDidClick(inAppNotification, data, keyValueMap, btnClickIndex);
         }
+    }
+
+    @SuppressLint("NewApi")
+    public void showHardPermissionPrompt() {
+        if (isPackageAndOsTargetsAbove(this, 32)) {
+            requestPermission();
+        }
+    }
+
+    @SuppressLint("NewApi")
+    public void showHardPermissionPrompt(CTInAppNotificationButton ctInAppNotificationButton){
+        if (isPackageAndOsTargetsAbove(this, 32)) {
+            if (ctInAppNotificationButton != null) {
+                isFallbackSettingsEnabled = ctInAppNotificationButton.isFallbackToSettings();
+            }
+            requestPermission();
+        }
+    }
+
+    @RequiresApi(api = 33)
+    public void requestPermission() {
+        int permissionStatus = ContextCompat.checkSelfPermission(InAppNotificationActivity.this,
+                Manifest.permission.POST_NOTIFICATIONS);
+
+        if (permissionStatus == PackageManager.PERMISSION_DENIED){
+            boolean isFirstTimeRequest = StorageHelper.getBoolean(InAppNotificationActivity.this,
+                    IS_FIRST_TIME_PERMISSION_REQUEST,true);
+            if (!isFirstTimeRequest) {
+                boolean neverAskAgainClicked = ActivityCompat.shouldShowRequestPermissionRationale(
+                        InAppNotificationActivity.this, ANDROID_PERMISSION_STRING);
+
+                if (shouldShowFallbackAlertDialog(neverAskAgainClicked)){
+                    showFallbackAlertDialog();
+                    return;
+                }
+
+                permissionCallbackWeakReference.get().onReject();
+                didDismiss(null);
+                return;
+            }
+
+            ActivityCompat.requestPermissions(InAppNotificationActivity.this,
+                    new String[]{ANDROID_PERMISSION_STRING}, PERMISSION_REQUEST_CODE);
+        }else{
+            permissionCallbackWeakReference.get().onAccept();
+            didDismiss(null);
+        }
+    }
+
+    /**
+     * This method will show `showFallbackAlertDialog()` if any of the below conditions are satisfied
+     * 1)If `neverAskAgainClicked` is true and when `isFbSettings` true.
+     *  `isFbSettings` key is available from IAM campaign.
+     * 2)If `neverAskAgainClicked` is true and when `showFbSettings` is true.
+     *  `showFbSettings` key is available when hard push permission flow is called.
+     * 3)If `neverAskAgainClicked` is true and when `inAppNotification.fallBackToNotificationSettings()` is true.
+     * `inAppNotification.fallBackToNotificationSettings()` is available when push primer flow is called.
+     * @param neverAskAgainClicked - This boolean will be true when permission is already denied.
+     *
+     * @return true/false
+     */
+    private boolean shouldShowFallbackAlertDialog(boolean neverAskAgainClicked) {
+        if (neverAskAgainClicked && isFallbackSettingsEnabled){
+            return true;
+        }else if (neverAskAgainClicked && shouldShowFallbackSettings){
+            return true;
+        }else return neverAskAgainClicked && inAppNotification.fallBackToNotificationSettings();
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        StorageHelper.putBoolean(InAppNotificationActivity.this,IS_FIRST_TIME_PERMISSION_REQUEST,
+                false);
+        if (requestCode == PERMISSION_REQUEST_CODE) {
+            boolean granted = grantResults.length > 0 && grantResults[0] ==
+                    PackageManager.PERMISSION_GRANTED;
+            if (granted) {
+                permissionCallbackWeakReference.get().onAccept();
+            }else {
+                permissionCallbackWeakReference.get().onReject();
+            }
+            didDismiss(null);
+        }
+    }
+
+    public void showFallbackAlertDialog() {
+        AlertDialogPromptForSettings.show(this, () -> {
+            Utils.navigateToAndroidSettingsForNotifications(InAppNotificationActivity.this);
+            didDismiss(null);
+            return Unit.INSTANCE;
+        }, () -> {
+            didDismiss(null);
+            return Unit.INSTANCE;
+        });
     }
 
     void didDismiss(Bundle data) {
@@ -152,7 +292,7 @@ public final class InAppNotificationActivity extends FragmentActivity implements
         }
         finish();
         InAppListener listener = getListener();
-        if (listener != null && getBaseContext() != null) {
+        if (listener != null && getBaseContext() != null && inAppNotification != null) {
             listener.inAppNotificationDidDismiss(getBaseContext(), inAppNotification, data);
         }
     }
@@ -190,6 +330,10 @@ public final class InAppNotificationActivity extends FragmentActivity implements
 
     void setListener(InAppListener listener) {
         listenerWeakReference = new WeakReference<>(listener);
+    }
+
+    public void setPermissionCallback(PermissionCallback callback) {
+        permissionCallbackWeakReference = new WeakReference<>(callback);
     }
 
     private CTInAppBaseFullFragment createContentFragment() {
@@ -250,13 +394,26 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                                         inAppNotification.getCampaignId());
                                                 data.putString("wzrk_c2a",
                                                         inAppNotification.getButtons().get(0).getText());
-                                                didClick(data, null);
+                                                didClick(data, null, 0);
                                                 String actionUrl = inAppNotification.getButtons().get(0)
                                                         .getActionUrl();
                                                 if (actionUrl != null) {
                                                     fireUrlThroughIntent(actionUrl, data);
                                                     return;
                                                 }
+                                                if (inAppNotification.isLocalInApp()) {
+                                                    showHardPermissionPrompt();
+                                                    return;
+                                                }
+
+                                                if (inAppNotification.getButtons().get(0).getType() != null &&
+                                                        inAppNotification.getButtons().get(0).getType()
+                                                        .equalsIgnoreCase(Constants.KEY_REQUEST_FOR_NOTIFICATION_PERMISSION)){
+                                                    showHardPermissionPrompt(
+                                                            inAppNotification.getButtons().get(0));
+                                                    return;
+                                                }
+
                                                 didDismiss(data);
                                             }
                                         })
@@ -272,12 +429,20 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                                     inAppNotification.getCampaignId());
                                             data.putString("wzrk_c2a",
                                                     inAppNotification.getButtons().get(1).getText());
-                                            didClick(data, null);
+                                            didClick(data, null, 1);
                                             String actionUrl = inAppNotification.getButtons().get(1).getActionUrl();
                                             if (actionUrl != null) {
                                                 fireUrlThroughIntent(actionUrl, data);
                                                 return;
                                             }
+
+                                            if (inAppNotification.getButtons().get(1).getType() != null &&
+                                                    inAppNotification.getButtons().get(1).getType()
+                                                    .equalsIgnoreCase(Constants.KEY_REQUEST_FOR_NOTIFICATION_PERMISSION)){
+                                                showHardPermissionPrompt(inAppNotification.getButtons().get(1));
+                                                return;
+                                            }
+
                                             didDismiss(data);
                                         }
                                     });
@@ -296,7 +461,7 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                                         inAppNotification.getCampaignId());
                                                 data.putString("wzrk_c2a",
                                                         inAppNotification.getButtons().get(0).getText());
-                                                didClick(data, null);
+                                                didClick(data, null, 0);
                                                 String actionUrl = inAppNotification.getButtons().get(0)
                                                         .getActionUrl();
                                                 if (actionUrl != null) {
@@ -304,6 +469,7 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                                     return;
                                                 }
                                                 didDismiss(data);
+
                                             }
                                         }).create();
                         if (inAppNotification.getButtons().size() == 2) {
@@ -317,7 +483,7 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                                     inAppNotification.getCampaignId());
                                             data.putString("wzrk_c2a",
                                                     inAppNotification.getButtons().get(1).getText());
-                                            didClick(data, null);
+                                            didClick(data, null, 1);
                                             String actionUrl = inAppNotification.getButtons().get(1).getActionUrl();
                                             if (actionUrl != null) {
                                                 fireUrlThroughIntent(actionUrl, data);
@@ -339,7 +505,7 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                         data.putString(Constants.NOTIFICATION_ID_TAG,
                                                 inAppNotification.getCampaignId());
                                         data.putString("wzrk_c2a", inAppNotification.getButtons().get(2).getText());
-                                        didClick(data, null);
+                                        didClick(data, null, 2);
                                         String actionUrl = inAppNotification.getButtons().get(2).getActionUrl();
                                         if (actionUrl != null) {
                                             fireUrlThroughIntent(actionUrl, data);
@@ -350,12 +516,13 @@ public final class InAppNotificationActivity extends FragmentActivity implements
                                 });
                     }
                 }
-                if(alertDialog != null){
+                if (alertDialog != null) {
                     alertDialog.show();
                     isAlertVisible = true;
                     didShow(null);
-                }else{
-                    config.getLogger().debug("InAppNotificationActivity: Alert Dialog is null, not showing Alert InApp");
+                } else {
+                    config.getLogger()
+                            .debug("InAppNotificationActivity: Alert Dialog is null, not showing Alert InApp");
                 }
                 break;
             }